Trimmed Repetitive Step Defs
