+++ src/test/java/com/library2/step_definitions/US07_AddUsers_StepDefs.java
@@ -18,17 +18,14 @@ public class US07_AddUsers_StepDefs {
     LoginPage loginPage = new LoginPage();
     US07_AddUsers_Page addUserPage = new US07_AddUsers_Page();
 
-    @Given("I am on the login page")
-    public void i_am_on_the_login_page() {
 
-    }
     @Given("I login using following credentials:")
     public void i_login_using_following_credentials(Map<String, String> credentials) {
         loginPage.userName.sendKeys(credentials.get("email"));
         loginPage.password.sendKeys(credentials.get("password"));
         loginPage.submit.click();
     }
-    @Given("I click on {string} module")
+    @Given("I click on {string}  module")
     public void i_click_on_module(String page) {
         String locator = "//span[text() ='"+page+ "']";
         Driver.getDriver().findElement(By.xpath(locator)).click();
