+++ src/test/java/com/library2/step_definitions/UserLandingPage_StepDefs.java
@@ -11,11 +11,8 @@ public class UserLandingPage_StepDefs {
 
     LoginPage loginPage = new LoginPage();
 
-    @Given("I am on the login page")
-    public void i_am_on_the_login_page() {
-        Driver.getDriver().get(ConfigurationReader.getProperty("url"));
-    }
-    @When("I login using {string} and {string}")
+
+    @When("I login using {string} and {string}.")
     public void i_login_using_and_library_user(String username, String password) {
         loginPage.login(username, password);
     }
